Move error display attributes into translation yml
This is the first part of decoupling error display/rendering options
from the code. This sets us up to move error dispaly options to live
with the error definitions in i18n/errors/en.yml.

A new class ErrorTranslation encapsulates error messages translations,
providing access to both the text and the display attributes.

Note that this commit does not convert any messages or remove the
various convenience exceptions (such as ErrorNoLog) - that will follow
this work shortly.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>