Handle single failed job when creating cookbook or policyfile
When I updated the local work (creating temp cookbook and policyfile
bundle) to use the job framework, I neglected to handle what happens
if the job fails. Exceptions would get thrown inside the job correctly
and the error reporter would print it, but that would not be visible to
the user since the job failure was not handled and the spinners would
overwrite the failure message. This fixes that to correctly raise an
error if there is a failed job.

Signed-off-by: tyler-ball <tball@chef.io>