Revert https://github.com/chef/chef-apply/pull/96
I mistakenly thought that telemetry data would be sent even if telemetry
is disabled. But there is already a check in the `run` method of
sender.rb that prevents this. Reverting the above PR because it is
unnecessary and prevents cleaning up the telemetry data files (which
happens regardless of whether they are sent successfully).

This reverts commit 6beb5a73e06349af5aa2a8962967f4788dbc0bf1.