Slim the gem down by removing specs + switch from train to train-core
Also use the new test script better caches the gems during the run.

Remove the specs and the readme from the gem artifact to reduce the size
of the install artifact. This drops our Workstation install size by 51
files and 238k on disk.

I changed the train dep from train to train-core since we also dep on
chef 15+ which includes train-core as a dep. There's no reason to
install both versions of train in CI for testing. In the workstation
package we'll end up installing both, but this is *more* correct. We
don't need all the azure/gce/ec2 deps just to test chef-apply since it
doesn't use any of that.

Signed-off-by: Tim Smith <tsmith@chef.io>