Disable strict config checking
Strict mode was originally enabled to provide a better UX
around validated configuration entries, the scope of chef-workstation
configuration has expanded beyond this single tool.

Disabling this allows us to avoid adding config support to chef-apply
every time we add new config keys for chef workstation app,
chef command line, chef-analyze, etc.  The trade-off is that
we will not be able to detect when the operator has a bad configuration
that includes keys that are genuinely invalid.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>