Add missing pass-through for `capture`
When we changed over to using Chef::Telemeter.timed_run_capture, it calls `capture`
This works in the normal case, but in error handling this routes back to
bthe ChefApply::Telemeter - which doesn't have a `capture`.

This would cause us to fail with a telemtry capture error during
standard error handling.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>